Move files out of subdirectory and rename local repository.
All the files were moved out of `asgmt3/` subfolder into the root directory of
the repository, and the local copy of the repository was renamed to reflect the
fact that the repository only holds code and test data for assignment three,
rather than anything and everything for the course.

Additionally, the
[`coveralls-maven-plugin`](https://github.com/hazendaz/coveralls-maven-plugin)
was added to the `pom.xml`.